+++ src/Farmer/Arm/DocumentDb.fs
@@ -2,7 +2,6 @@
 module Farmer.Arm.DocumentDb
 
 open Farmer
-open Farmer.CoreTypes
 open Farmer.CosmosDb
 
 let containers = ResourceType ("Microsoft.DocumentDb/databaseAccounts/sqlDatabases/containers", "2020-03-01")
@@ -28,9 +27,9 @@ module DatabaseAccounts =
                 |}
             }
             interface IArmResource with
-                member this.ResourceName = this.Name
+                member this.ResourceId = containers.resourceId (this.Account/this.Database/this.Name)
                 member this.JsonModel =
-                    {| containers.Create(this.Account/this.Database/this.Name, dependsOn = [ ResourceId.create this.Database ])
+                    {| containers.Create(this.Account/this.Database/this.Name, dependsOn = [ sqlDatabases.resourceId this.Database ])
                         with
                            properties =
                                {| resource =
@@ -68,9 +67,9 @@ module DatabaseAccounts =
           Account : ResourceName
           Throughput : int<RU> }
         interface IArmResource with
-            member this.ResourceName = this.Name
+            member this.ResourceId = sqlDatabases.resourceId (this.Account/this.Name)
             member this.JsonModel =
-                {| sqlDatabases.Create(this.Account/this.Name, dependsOn = [ ResourceId.create this.Account ]) with
+                {| sqlDatabases.Create(this.Account/this.Name, dependsOn = [ accounts.resourceId this.Account ]) with
                        properties =
                            {| resource = {| id = this.Name.Value |}
                               options = {| throughput = string this.Throughput |} |}
@@ -111,7 +110,7 @@ type DatabaseAccount =
     ]
 
     interface IArmResource with
-        member this.ResourceName = this.Name
+        member this.ResourceId = databaseAccounts.resourceId this.Name
         member this.JsonModel =
             {| databaseAccounts.Create(this.Name, this.Location, tags = this.Tags) with
                    kind = "GlobalDocumentDB"