+++ RELEASE_NOTES.md
@@ -1,13 +1,17 @@
 Release Notes
 =============
 ## 1.3.0-beta1
-* Container Instance: Support for command line arguments
-* Key Vault: Support for setting tags on key vault secrets
+* Container Instance: Support for command line arguments.
+* Key Vault: Support for setting tags on key vault secrets.
+* Deployment Scripts: Run Azure CLI commands as part of an ARM deployment (PowerShell or AzCli).
 
 * Storage Account: Support for the full set of Storage Account Kind and SKUs (minor breaking change).
 * Storage Account: Improved integration with CDN.
 
-* CDN has improved integration with Storage Accounts.
+* CDN: Improved integration with Storage Accounts.
+* CDN: Fixed an issue where custom domains did not generate the correct ARM name path.
+
+* ARM generation: Smarter emitting of "raw" ARM expressions.
 
 ## 1.2.0
 * Log Analytics: Initial release.